+++ libraries/imu.h
@@ -1,18 +1,20 @@
 #pragma once //(or use a guard symbol)
 
+#include <stdint.h>
+
 class IMU {
   public:
     virtual float pitch(); // degrees
     virtual float yaw(); // degrees
     virtual float roll(); // degrees
 
-    virtual float accelerationX(); // m/s^2
-    virtual float accelerationY(); // m/s^2
-    virtual float accelerationZ(); // m/s^2
+    virtual int16_t accelerationX(); // cm/s^2
+    virtual int16_t accelerationY(); // cm/s^2
+    virtual int16_t accelerationZ(); // cm/s^2
 
-    virtual float gyroX();
-    virtual float gyroY();
-    virtual float gyroZ();
+    virtual int16_t gyroX(); // degrees/s
+    virtual int16_t gyroY(); // degrees/s
+    virtual int16_t gyroZ(); // degrees/s
 
     virtual void setup();
     virtual void calibrate();
