+++ libraries/lsm9ds1_api.h
@@ -1,6 +1,7 @@
 #include <Adafruit_LSM9DS1.h>
 #include <Adafruit_Sensor.h>
 #include "imu.h"
+#include <stdint.h>
 
 class LSM9DS1_API: public IMU {
   static Adafruit_LSM9DS1 sensor;
@@ -11,13 +12,13 @@ class LSM9DS1_API: public IMU {
     float yaw();
     float roll();
 
-    float accelerationX();
-    float accelerationY();
-    float accelerationZ();
+    int16_t accelerationX();
+    int16_t accelerationY();
+    int16_t accelerationZ();
 
-    float gyroX();
-    float gyroY();
-    float gyroZ();
+    int16_t gyroX();
+    int16_t gyroY();
+    int16_t gyroZ();
 
     float magX();
     float magY();
